refactor: use latest terminology
