+++ test/fixtures/historyapifallback-3-config/index.html
@@ -1,1 +1,1 @@
-contentBase file
+static file
