+++ test/fixtures/static-config/foo.js
