+++ test/fixtures/static-config/other/foo.html
