+++ test/fixtures/static-config/public/bar/index.html
