+++ test/fixtures/static-config/public/index.html
