+++ test/fixtures/static-config/public/other.html
