+++ test/fixtures/static-config/static/index.html
