+++ test/fixtures/static-config/webpack.config.js
