+++ test/server/static-directory-option.test.js
@@ -6,13 +6,10 @@ const webpack = require("webpack");
 const request = require("supertest");
 const Server = require("../../lib/Server");
 const testServer = require("../helpers/test-server");
-const config = require("../fixtures/contentbase-config/webpack.config");
+const config = require("../fixtures/static-config/webpack.config");
 const port = require("../ports-map")["static-directory-option"];
 
-const staticDirectory = path.resolve(
-  __dirname,
-  "../fixtures/contentbase-config"
-);
+const staticDirectory = path.resolve(__dirname, "../fixtures/static-config");
 const publicDirectory = path.resolve(staticDirectory, "public");
 const otherPublicDirectory = path.resolve(staticDirectory, "other");
